Extract into separate files
